+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/model/ReceiverError.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.share.receive.model
+
+enum class ReceiverError {
+    NO_INTERNET_CONNECTION,
+    CANT_CONNECT_TO_SERVER,
+    INVALID_FILE_FORMAT,
+    EXPIRED_LINK
+}