+++ CHANGELOG.md
@@ -1,19 +1,27 @@
 # Changelog
 
-# 1.3.1 - In progress
+# 1.4.1 - In Progress
 
+# 1.4.0
+
+- [Feature] Placeholder view on receive key
 - [Feature] Add forbidden frequency dialog
 - [Feature] Rework all system dialog to custom
-- [Feature] New share flow(with `sf#path=()&id=()key=`)
+- [Feature] New share flow(with `sf#path=()&id=()key=`) + metric + error handling
 - [Feature] Share shadow file if that exist on NFC
 - [Feature] Add application catalog button
 - [Feature] Add application catalog list with sorted button
 - [Feature] Add categories, categories list and search
 - [Feature] Add installed, new switch, install button
+- [Feature] Add support for 3 version of NFC file
 - [CI] Bump deps
 - [FIX] Uploading share errors
 - [FIX] Display by white color meta data on NFC card
 - [FIX] Scrim status bar and design changes on bottom sheet
+- [FIX] Fix crash when delete from trash key with favorites
+- [CI] Git submodule `Flipper Protobuf`
+- [FIX] Favorites synchronization
+- [CI] Fix local build of application
 
 # 1.3.0